requirements/local.txt
# Include base requirements
-r base.txt
-r test.txt  # Include test dependencies for local development

# Development tools
gunicorn==23.0.0  # WSGI server (same as production for consistency)
django-debug-toolbar==6.1.0
ipython==8.31.0  # Enhanced Python shell
django-extensions==3.2.3  # Additional management commands

# Code quality (using Ruff instead of flake8, black, isort)
ruff==0.7.4
ruff-lsp==0.0.53

# Environment management
python-dotenv==1.0.1  # Load .env files
